requirements.txt
streamlit>=1.25.0
diffusers>=0.21.0
torch>=1.13.0
transformers>=4.21.0
pillow>=9.0.0
accelerate>=0.21.0